remove namespace from resource spec
