+++ pkg/policyviolation/helpers.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/golang/glog"
 	kyverno "github.com/nirmata/kyverno/pkg/api/kyverno/v1"
+	client "github.com/nirmata/kyverno/pkg/dclient"
 	v1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	unstructured "k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -30,21 +31,6 @@ func converLabelToSelector(labelMap map[string]string) (labels.Selector, error) 
 	return policyViolationSelector, nil
 }
 
-func containsOwner(owners []kyverno.ResourceSpec, pvResourceSpec kyverno.ResourceSpec) bool {
-	curOwner := kyverno.ResourceSpec{
-		Kind:      pvResourceSpec.Kind,
-		Namespace: pvResourceSpec.Namespace,
-		Name:      pvResourceSpec.Name,
-	}
-
-	for _, targetOwner := range owners {
-		if reflect.DeepEqual(curOwner, targetOwner) {
-			return true
-		}
-	}
-	return false
-}
-
 // validDependantForDeployment checks if resource (pod) matches the intent of the given deployment
 // explicitly handles deployment-replicaset-pod relationship
 func validDependantForDeployment(client appsv1.AppsV1Interface, pvResourceSpec kyverno.ResourceSpec, resource unstructured.Unstructured) bool {
@@ -58,15 +44,14 @@ func validDependantForDeployment(client appsv1.AppsV1Interface, pvResourceSpec k
 	}
 
 	owner := kyverno.ResourceSpec{
-		Kind:      pvResourceSpec.Kind,
-		Namespace: pvResourceSpec.Namespace,
-		Name:      pvResourceSpec.Name,
+		Kind: pvResourceSpec.Kind,
+		Name: pvResourceSpec.Name,
 	}
 
 	start := time.Now()
-	deploy, err := client.Deployments(owner.Namespace).Get(owner.Name, metav1.GetOptions{})
+	deploy, err := client.Deployments(resource.GetNamespace()).Get(owner.Name, metav1.GetOptions{})
 	if err != nil {
-		glog.Errorf("failed to get resourceOwner deployment %s/%s/%s: %v", owner.Kind, owner.Namespace, owner.Name, err)
+		glog.Errorf("failed to get resourceOwner deployment %s/%s/%s: %v", owner.Kind, resource.GetNamespace(), owner.Name, err)
 		return false
 	}
 	glog.V(4).Infof("Time getting deployment %v", time.Since(start))
@@ -74,12 +59,12 @@ func validDependantForDeployment(client appsv1.AppsV1Interface, pvResourceSpec k
 	// TODO(shuting): replace typed client AppsV1Interface
 	expectReplicaset, err := deployutil.GetNewReplicaSet(deploy, client)
 	if err != nil {
-		glog.Errorf("failed to get replicaset owned by %s/%s/%s: %v", owner.Kind, owner.Namespace, owner.Name, err)
+		glog.Errorf("failed to get replicaset owned by %s/%s/%s: %v", owner.Kind, resource.GetNamespace(), owner.Name, err)
 		return false
 	}
 
 	if reflect.DeepEqual(expectReplicaset, v1.ReplicaSet{}) {
-		glog.V(2).Infof("no replicaset found for deploy %s/%s/%s", owner.Namespace, owner.Kind, owner.Name)
+		glog.V(2).Infof("no replicaset found for deploy %s/%s/%s", resource.GetNamespace(), owner.Kind, owner.Name)
 		return false
 	}
 	var actualReplicaset *v1.ReplicaSet
@@ -104,4 +89,36 @@ func validDependantForDeployment(client appsv1.AppsV1Interface, pvResourceSpec k
 		}
 	}
 	return false
+}
+
+// GetOwner of a resource by iterating over ownerReferences
+func GetOwner(dclient *client.Client, ownerMap map[kyverno.ResourceSpec]interface{}, resource unstructured.Unstructured) {
+	var emptyInterface interface{}
+	resourceSpec := kyverno.ResourceSpec{
+		Kind: resource.GetKind(),
+		Name: resource.GetName(),
+	}
+	if _, ok := ownerMap[resourceSpec]; ok {
+		// owner seen before
+		// breaking loop
+		return
+	}
+	rOwners := resource.GetOwnerReferences()
+	// if there are no resource owners then its top level resource
+	if len(rOwners) == 0 {
+		// add resource to map
+		ownerMap[resourceSpec] = emptyInterface
+		return
+	}
+	for _, rOwner := range rOwners {
+		// lookup resource via client
+		// owner has to be in same namespace
+		owner, err := dclient.GetResource(rOwner.Kind, resource.GetNamespace(), rOwner.Name)
+		if err != nil {
+			glog.Errorf("Failed to get resource owner for %s/%s/%s, err: %v", rOwner.Kind, resource.GetNamespace(), rOwner.Name, err)
+			// as we want to process other owners
+			continue
+		}
+		GetOwner(dclient, ownerMap, *owner)
+	}
 }
